fix learninig rate scheduler

+++ src/polimi/utils/tf_models/utils/lr_schedule.py
@@ -7,6 +7,6 @@ def get_simple_decay_scheduler(scheduling_rate):
         if epoch < 5:
             return lr
         else:
-            return lr * tf.math.exp(-scheduling_rate)
+            return float(lr * tf.math.exp(-scheduling_rate))
         
     return tf.keras.callbacks.LearningRateScheduler(scheduler)
